requirements.txt
Django==3.2.4
djangorestframework==3.12.4
dj-rest-auth==2.1.11
django-allauth==0.45.0
